docs(projects): add Ace by DAISY to the projects list
Ace (as in “Accessibility Checker for EPUB”) is a tool developed by
the DAISY Consortium to run automated checks and content extraction
in order to assist with the evaluation of EPUB accessibility.

It uses aXe under the hood to check EPUB’s XHTML Content Documents.

More info at https://daisy.github.io/ace/