fix(rule): Remove coverage of svg[role="img"] for image-alt rule
We have been advised that best practice is to use the following for inline SVGs:
```
<svg role="img">
  <title>SVG Title</title>
</svg>
```
which is echoed in [this article](https://css-tricks.com/accessible-svgs/#article-header-id-6).

I see that there is conversation in this issue https://github.com/dequelabs/axe-core/issues/20 about how to best go about it, but for now the test for `svg[role="img"]` needs to be removed since it isn't really valid.