doc: update rule-descriptions
This patch updates the `rule-descriptions.md` document via `npx grunt build`. It appears the source files in which this document have changed but the generated content has not been comitted.

I've been "ignoring" this change in the many PRs I've opened recently and have finally grown tired of doing so.