fix(imports): try/catch with fallback to avoid CSP issues
This patch updates the logic we use to get a reference to the global scope. Previously, if a page (or Chrome extension) had a strict CSP set ("unsafe-eval"), we caused an exception to be thrown by using the `Function` constructor. Now we `try/catch` the statement and fallback to using `window`.