fix: check for required functionality before running an analysis
This patch adds a few feature checks to the axe-core library. This ensures axe-core is able to rely on native methods that the host environment may not have (or may have overwritten with broken code).

If the host environment is missing functionality, a helpful "SupportError" will be raised rather than an unexpected crash.

I'm not sure if I've added these feature tests in the correct place. Happy to move them if need be :)

Closes #1136.