chore: add grunt configure to run on pre-commit hook (#1244)
**Short Description:**
Preventing us to open PR's like this - https://github.com/dequelabs/axe-core/pull/1184

**Long Description:**
There has been several PR's in the past, where changes are done, but the `rule-descriptions` are not updated, because `build` was not run as a part of the development efforts.

These usually happen when the metadata of a rule spec or check spec is changed. 

Overhead being unrelated changes show up on other work.

This PR adds `grunt configure`, which builds `doc/rule-descriptions.md` as a part of pre-commit hook. This should alleviate the above 

Closes issue: NA

## Reviewer checks

**Required fields, to be filled out by PR reviewer(s)**
- [x] Follows the commit message policy, appropriate for next version
- [x] Has documentation updated, a DU ticket, or requires no documentation change
- [x] Includes new tests, or was unnecessary
- [x] Code is reviewed for security by: @WilcoFiers