fix: prevent TypeErrors in color-contrast checks
This patch adds a safeguard to `color.filteredRectStack` which prevents a `TypeError: Cannot read property '0' of undefined` from being thrown.

This is not a perfect solution, but a stop-gap instead. It's still unclear why we're getting `undefined` here. More investigation is needed, but this can happen at a later date.

Closes #1306 and #1259.