chore: disable "max-statements" and "complexity" ESLint rules
This patch removes the arbitrary and ignored "max-statements" and "complexity" ESLint rules. Though the intention of having small functions with a low cyclomatic complexity is great, we have not followed these rules in practice. Instead, we've just repeatedly increased the thresholds when necessary. Doing so just adds noise for the reader of the code sift through, and is a tedious chore for the author of the code to have to do.