test: make the output of `test:examples` readable
This patch rewrites the program which runs the examples tests with something more human (and coincidentally machine) readable. Rather than piping `npm install` and `npm test` output to `stdio` directly (and in parallel), we suppress the output (unless an error is raised) and run the tests serially. We give visual feedback letting the user running the tests know what's happening, but don't flood them with too much information (this is done via [`Listr`](http://npmjs.org/listr)).

This has the unfortunate side effect of increasing the time the `test:examples` script takes, but will make debugging the tests significantly easier.