perf(core/matchesExpression): avoid allocating memory
`matchesExpression` is a recursive function that processes the
`expressions` array one item at a time (`exps.pop()`). It currently
allocates a new array for each recursive call, which is unnecessary.

This patch introduces another parameter (`index`) to keep track of the
expression to match on, thus avoiding the memory allocation. The bulk
of the logic is moved to another function (`optimizedMatchesExpression`)
to keep the same parameters for the exported function
(`matchesExpression`).

Benchmarking with `hyperfine --warmup 1 'axe https://www.dr.dk/'` shows
a reduction in running time of ~400ms.