perf(core): reduce memory allocations (#3053)
* perf(core/virtual-node): cache node properties

`props` is being accessed frequently. Returning a new object for every
access results in unnecessary overhead due to the memory allocation and
the subsequent garbage collection.

Tests that relied on mutating private fields have been updated.

Benchmarking with `hyperfine --warmup 1 'axe https://www.dr.dk/'` shows
a reduction in running time of ~400ms.

* perf(core/matchExpressions): avoid array allocation inside hot loop

There is no need to allocate a new array to iterate over the members of
two arrays.

This patch switches the iteration logic to use an index for reading from
both arrays. Added a test for this code-path.

* perf(core/matchesExpression): avoid allocating memory

`matchesExpression` is a recursive function that processes the
`expressions` array one item at a time (`exps.pop()`). It currently
allocates a new array for each recursive call, which is unnecessary.

This patch introduces another parameter (`index`) to keep track of the
expression to match on, thus avoiding the memory allocation. The bulk
of the logic is moved to another function (`optimizedMatchesExpression`)
to keep the same parameters for the exported function
(`matchesExpression`).

Benchmarking with `hyperfine --warmup 1 'axe https://www.dr.dk/'` shows
a reduction in running time of ~400ms.

* perf(core/matchExpressions): avoid allocating memory

There is no need to copy and reverse `vNodes` - an index can be used to
keep track of the current node.

* perf(core/matchExpressions): recycle local variables

Avoid allocating new object for during every call to
`createLocalVariables`. Sharing `recycledLocalVariables` between calls
to `matchExpressions` is safe due to the fact that `axe-core` runs in
a single thread.

Benchmarking with
`hyperfine --warmup 1 'axe https://www.theguardian.com/international'`
shows a reduction in running time of ~80ms.

* perf(core/matchExpressions): avoid array allocations

Initialize `childOnly` and `childAny` with `null` and only create new
arrays when adding data for the first time.

Co-authored-by: Sorin Davidoi <sorin.davidoi@falcon.io>