Add a rule to check if a page has a sign language widget
This rule is implemented currently for the Mind Rockets Sign Language Widget, but it could also be extended in the future for other automated Sign Language widgets.
In addition, it checks if there are any tags in the page which detects the page has anything related to deaf or sign language.