feat(i18n): Add Italian translation (#4344)
create it.json following template

This pull request contains the Italian translation of a JSON template
related to accessibility auditing. The translation includes all the keys
and values of the original JSON, ensuring that the meaning and context
of each message are accurately conveyed in Italian. The translation was
carefully reviewed to ensure that it meets quality standards