fix: correct typos in texts (#4499)
Hi! I noticed some minor typos and inconsistent text capitalization, as
well as what looks like flipped `help` and `description` text in one
spot for the `landmark-unique` rule. This PR fixes the ones I was able
to spot.

I noticed some overlap with
https://github.com/dequelabs/axe-core/issues/4385, I'd be happy to grab
the text changes mentioned there in this PR as well, ("Ensures" ->
"Ensure" in rule descriptions), if that's not stepping on anybody's
toes.