refactor: put isContext methods on axe.utils (#4524)
Had a couple scenarios where this code was getting copied over to other
repos. For backward compat reasons that's going to need to stay, but at
least if we have these methods on axe.utils we can prevent problems in
the future if we change how context works.

No QA needed on this one.