docs: fix example tests (#4564)
When I updated https://github.com/dequelabs/axe-core/pull/4548 to latest
it caused the qunit tests to fail. Turns out that
[qunitjs](https://www.npmjs.com/package/qunitjs) has been deprecated and
is now called just `qunit`, and `grunt-contrib-qunit` [has
issues](https://github.com/gruntjs/grunt-contrib-qunit/issues/209) if
not using latest qunit. Once upgraded I was able to get the tests
passing but only in headed mode. That was because puppeteer [is now
headless by
default](https://github.com/gruntjs/grunt-contrib-qunit/blob/main/docs/qunit-options.md#puppeteer),
so passing the `ignoreDefaultArgs` param and the `--headless` arg was
causing the issue.

`jest_react` was also failing and after trying to fix it with it's
current dependencies @dbjorge suggested it would be better to just
change it over to `@react/testing-library`. It also means we can support
react 18 in the examples too.