test: ignore windowHeight in partialRun tests (#4571)
This should fix the issue with the tests failing on`windowHeight` being
different between two consecutive runs.

I also added a deep axe-core results compare function that can do two
significant things for us when comparing result objects. First it can
ignore keys (like `timestamp`) so we don't have to always set it equal
to each other in the tests. Second is that instead of dumping the entire
result objects when a value is different this will output they key path
and the values that are different which make debugging a lot easier.

```
Expected result.testEnvironment.windowHeight to equal "948" but got "896"
```