docs: Update README.md with grammatical fixes (#4583)
Update the doc for clarity and consistency with present tense. Also,
"rules' description and help properties as well as each checks'
metadata.messages property" → should be "rule's description and help
properties, as well as each check's metadata.messages property"
(singular possessive for "rule" and "check," and add a comma before "as
well as").