test(preload): fix link to gutenberg css (#4730)
Seems the unpkg link `https://unpkg.com/gutenberg-css@0.4` now redirects
to a `.scss` file rather than a css file, which caused the link to fail
to load the resource. Updated the link to hardcode the css file path.

Also added a `return` to the `done` call as it would call done on
failure, but then continue to try to remove the stylesheet that didn't
exist.