ci(nightly): port testing to github actions (#4918)
This patch begins moving the workflows from CircleCI to Github. In order
to minimize the reviews, the first chunk is only converting the nightly
tests.

# Key Changes

## Removal of Circle CI Config for Nightly Tests

Since this patch adds support for running nightly tests in GitHub
Actions, we no longer need it in Circle CI. It would be wasteful of
resources to do the same work in multiple places.

## Install Deps Composite Action

This file reduces duplication of steps between multiple jobs and
workflows. It is called after a repository is checked out in CI in order
to install core OS dependencies, browsers, package dependencies, and
starts (when needed) a virtual X server for running headed browsers
against for testing.

## Nightly Tests

The nightly workflow runs the current beta releases for Firefox and
Chrome, WCAG ACT's current version, and the latest aria practices tests.

## Webdriver Configuration

Previously, `browser-driver-manager` was used to force install the
correct browser, driver, and updated selenium to match. This caused
specific quirks in ordering of steps that had to be adhered to. With
this conversion, the sources are now setup to take in the needed paths
by an environment variable. This allows the `install-deps` action to
install the browsers as they are an OS level dependency to run. Then the
paths are passed to the test runs and utilized. Bypassing the need to
run certain installs in a specific order.

# Verification

The original code in the PR was running on every push to get a test run
going. [This
run](https://github.com/dequelabs/axe-core/actions/runs/18976324783) was
from the commit before the final push which turned off the `on: push`
execution. This shows the workflow successfully running with all nightly
tests passing.

Refs: #4912