test: fix prototype integration test in dark mode (#4937)
This test case was failing when I ran it locally because it relied on
the default text `color` having low contrast with a `darkred`
background, which is true when the system color preference is light mode
but not in dark mode. This updates the test to also pin the foreground
color, making it agnostic to the system color mode.

No QA required