ci(tests): port to github actions (#4919)
This patch ports the main test workflow over to GHA. This is not
removing Circle CI configuration as we need to convert required checks
over. As well as only removing that after the publishing is ported over
to GitHub.

One key change in the patterns from the CircleCI system, is the browser
tests and the examples have to do the build internally instead of using
the build artifact. I have not had the time to fully investigate what is
going on with the differences. I _believe_ it fails in re-use now
because of some slight difference in the way the caches were done before
vs how they are done in GitHub Actions.

As an aside, I realized in the previous nightly PR I did mess up one
aspect. Since the browsers are individual steps now, if the first
running fails it was blocking the second from proceeding. For that
situation, I added an instruction to always run the second browser. This
is good enough for now and we can further optimize later when we setup
better failure reporting.

Refs: #4912

---------

Co-authored-by: Dan Bjorge <dan.bjorge@deque.com>