ci(deploy): setup node internally (#4948)
Setting up Node through a composite workflow does not inherit the
environment from the host. Meaning it creates its own token which does
not have the correct deployment authorization.

This patch moves setting up node internal with no more use of the
composite action for deployment. Which, enables the correct token to be
used. It has a side effect of also saving us time in deploy by not
installing the browser dependencies that aren't needed for this
workflow.

Closes: #4912